+++ integration-tests/server/rust_server/src/types/arrays_test.rs
@@ -1,0 +1,8 @@
+// Created by Fen v0.5.0 at 16:09:11 on 2025-02-14
+// Do not manually modify this file as it is automatically generated
+
+use uuid::Uuid;
+
+pub type Input = Vec<isize>;
+
+pub type Output = Vec<Uuid>;