+++ integration-tests/server/rust_server/src/types/nested_structs_test.rs
@@ -1,0 +1,22 @@
+// Created by Fen v0.5.0 at 16:09:11 on 2025-02-14
+// Do not manually modify this file as it is automatically generated
+
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+
+pub type Output = Person;
+
+#[derive(Serialize, Deserialize, Debug, Clone, Eq, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct Person {
+    pub name: String,
+    pub birthday: DateTime<Utc>,
+    pub car: Car,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone, Eq, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct Car {
+    pub color: String,
+    pub year: isize,
+}